(#1182) code 钱包配置

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/wallet/bill_detail/BillDetailFragment.java
@@ -93,7 +93,7 @@ public class BillDetailFragment extends TSFragment {
 
         mBillStatus.setText(getString(status == 0 ? R.string.transaction_doing : (status == 1 ? R.string.transaction_success : R.string.transaction_fail)));
         String moneyStr = (status == 1 ? (action == 0 ? "- " : "+ ") : "") + String.format(Locale.getDefault(), getString(R.string.dynamic_send_toll_select_money_),
-                mBillDetailBean.getAmount());
+                (float)mBillDetailBean.getAmount());
         mTvMineMoney.setText(moneyStr);
         mBillUserContainer.setVisibility(is_user ? View.VISIBLE : View.GONE);
         mBillAccountContainer.setVisibility(is_user ? View.GONE : View.VISIBLE);
